add a switch operation as a type of mutation. makes learning faster
given a string, the switch operation picks two characters randomly
and switches them.

A lot of the time the algo makes no progress because it's just waiting
for a chance mutation to put a character it needs in the same spot,
even though it might already know what all the correct characters are
(because of the char frequency and common chars score). So,
the switch operation makes it likelier to find a better solution since
it doesn't have to wait for the correct character to mutate into the string
at the right spot, it will just pick one of the correct characters already
present and try putting it in a different spot